Menu principal FIXED
Implementada la nueva colocacion.
+
Arreglado los errores de posicionamiento provovados por las Animations
tal y como se explica aqui:
"MenuScene in AndEngine is a very useful feature, it’s an extension of
Scene, giving us the tools to make easy and nice looking menus.

But it does have a few quirks, in the example below I’m creating a
simple MenuScene with 3 items, startMenu, highScoreMenu and quitMenu,
they will be in the middle of the screen, centered. But what if we
wanted to change the position of an item, for example if we wanted to
put one button in the left side, another in the right side of the
screen, well, in MenuScene, you can only change the position of an item
after the buildAnimations() if you do it before, it will be
disregarded."